requirements.txt
openai==0.28.0
streamlit==1.29.0
docx==0.2.4
pandas==2.1.3
python-dotenv==1.0.0
streamlit==1.29.0